refactor: test 코드 클래스별로 분리
LottoTest를 LottoTest, LottoStoreTest, LottoMachineTest, PlayerTest, CommonUtilTest로 각각 분리
